resolve circular dependency thing (I think?)
