+++ src/testing/jest/test/jest-runner.spec.ts
@@ -1,5 +1,4 @@
 import type * as d from '@stencil/core/declarations';
-import { ConfigFlags } from '../../../cli/config-flags';
 import { getEmulateConfigs, includeTestFile } from '../jest-runner';
 
 describe('jest-runner', () => {
@@ -75,7 +74,7 @@ describe('jest-runner', () => {
     const testing: d.TestingConfig = {
       emulate: [{ device: 'anDroiD' }, { device: 'iphone' }],
     };
-    const flags: ConfigFlags = {
+    const flags: d.ConfigFlags = {
       emulate: 'Android',
     };
     const emulateConfigs = getEmulateConfigs(testing, flags);
@@ -87,7 +86,7 @@ describe('jest-runner', () => {
     const testing: d.TestingConfig = {
       emulate: [{ userAgent: 'Mozilla/Android' }, { userAgent: 'SomeUserAgent/iPhone X' }],
     };
-    const flags: ConfigFlags = {
+    const flags: d.ConfigFlags = {
       emulate: 'android',
     };
     const emulateConfigs = getEmulateConfigs(testing, flags);
@@ -99,7 +98,7 @@ describe('jest-runner', () => {
     const testing: d.TestingConfig = {
       emulate: [{ device: 'android' }, { device: 'iphone' }],
     };
-    const flags: ConfigFlags = {};
+    const flags: d.ConfigFlags = {};
     const emulateConfigs = getEmulateConfigs(testing, flags);
     expect(emulateConfigs).toHaveLength(2);
   });
